no-redeploy script
this avoids re-deploying individual contracts, however, it has no
concept of dependencies, so for example if you wanted to redeploy the
ERC20Faucet, you could delete the appropiate key in the JSON and have it
re-deployed, but that would leave you with a FaucetFactory that still
uses the old contract. This is not ideal, and makes me miss hardhat just
a little.